auth/requirements.txt
fastapi
uvicorn
dependency-injector
redis[hiredis]
pyjwt
pydantic-settings